falcon-pdf-ocr/requirements.txt
transformers>=4.46
torchvision>=0.20,<0.21
pillow>=10.0
einops>=0.8
PyMuPDF>=1.24
tqdm>=4.66
accelerate>=0.34
safetensors>=0.4
